Collect resources

+++ src/main/java/swcnoops/server/model/StorageItem.java
@@ -1,7 +1,7 @@
 package swcnoops.server.model;
 
 public class StorageItem {
-    public long amount;
-    public long capacity;
+    public int amount;
+    public int capacity;
     public int scale;
 }
